backend/requirements.txt
# DocuQuery AI - Backend Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1

# LangChain (text splitting)
langchain-text-splitters==0.3.4

# Embeddings
sentence-transformers==3.3.1
openai==1.58.1

# LLM - Google Gemini
google-generativeai==0.8.4

# Vector Database
pinecone-client==5.0.1

# Supabase
supabase==2.11.0

# Redis Caching
redis[hiredis]==5.2.1

# Document Classification
scikit-learn==1.6.1
joblib==1.4.2
numpy==2.2.1

# HTTP Client
httpx==0.28.1
